move version to about us section
